refactor(drivers): reorganize driver package
