+++ diode_measurement/drivers/brandbox.py
@@ -1,11 +1,11 @@
 from typing import Any
 
-from comet.driver.hephy.brandbox import BrandBox
+from comet.driver.hephy.brandbox import BrandBox as _BrandBox
 
-__all__ = ["BrandBoxDriver"]
+__all__ = ["BrandBox"]
 
 
-class BrandBoxDriver(BrandBox):
+class BrandBox(_BrandBox):
     def configure(self, options: dict[str, Any]) -> None:
         self.open_all_channels()
         channels = options.get("channels", [])
