+++ diode_measurement/view/resource.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from PySide6 import QtCore, QtWidgets
 
-from ..driver import driver_factory
+from ..drivers import driver_factory
 from ..utils import open_resource
 
 __all__ = ["ResourceWidget"]
